+++ lib/pages/home_tecnico/home_tecnico.dart
@@ -1,10 +1,10 @@
-import 'package:caderneta_campo_digital/components/EstateCard/estate_card.dart';
-import 'package:caderneta_campo_digital/components/PendencyCard/pendency_card.dart';
-import 'package:caderneta_campo_digital/components/Topbar/topbar.dart';
+import 'package:caderneta_campo_digital/components/estate_card.dart';
+import 'package:caderneta_campo_digital/components/topbar.dart';
+import 'package:caderneta_campo_digital/pages/home_tecnico/components/pendency_card.dart';
 import 'package:flutter/material.dart';
 
-class HomeTecnico extends StatelessWidget {
-  const HomeTecnico({Key? key}) : super(key: key);
+class HomeTecnicoPage extends StatelessWidget {
+  const HomeTecnicoPage({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     Size size = MediaQuery.of(context).size;
@@ -16,21 +16,25 @@ class HomeTecnico extends StatelessWidget {
           children: const <Widget>[
             PendencyCard(),
             EstateCard(
-                estateName: "Propriedade 1",
-                estateAddress: "ENDERECO",
-                qtdPlot: 5,),
+              estateName: "Propriedade 1",
+              estateAddress: "ENDERECO",
+              qtdPlot: 5,
+            ),
             EstateCard(
-                estateName: "Propriedade 1",
-                estateAddress: "ENDERECO",
-                qtdPlot: 5,),
+              estateName: "Propriedade 1",
+              estateAddress: "ENDERECO",
+              qtdPlot: 5,
+            ),
             EstateCard(
-                estateName: "Propriedade 1",
-                estateAddress: "ENDERECO",
-                qtdPlot: 5,),
+              estateName: "Propriedade 1",
+              estateAddress: "ENDERECO",
+              qtdPlot: 5,
+            ),
             EstateCard(
-                estateName: "Propriedade 1",
-                estateAddress: "ENDERECO",
-                qtdPlot: 5,),
+              estateName: "Propriedade 1",
+              estateAddress: "ENDERECO",
+              qtdPlot: 5,
+            ),
           ],
         ),
       ),
