+++ pubspec.yaml
@@ -52,8 +52,6 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
     - assets/
     - assets/login/
     - assets/fonts/
